+++ src-docs/com/wsscode/pathom/book/connect/getting_started.cljs
@@ -3,30 +3,25 @@
             [com.wsscode.pathom.connect :as pc]
             [com.wsscode.pathom.profile :as pp]))
 
-; initialize a dispatch function
-(defmulti resolver-fn pc/resolver-dispatch)
-; initialize indexes
-(def indexes (atom {}))
-
-; this creates a factory that will add a new method on the resolver-fn and add it to the indexes
-(def defresolver (pc/resolver-factory resolver-fn indexes))
-
 ; creating our first resolver
-(defresolver `latest-product
+(pc/defresolver latest-product [_ _]
   {::pc/output [{::latest-product [:product/id :product/title :product/price]}]}
-  (fn [_ _]
-    {::latest-product {:product/id    1
-                       :product/title "Acoustic Guitar"
-                       :product/price 199.99M}}))
+  {::latest-product {:product/id    1
+                     :product/title "Acoustic Guitar"
+                     :product/price 199.99M}})
 
 (def parser
-  (p/parser {::p/plugins
-             [(p/env-plugin
-                {::p/reader             [p/map-reader
-                                         pc/all-readers]
-                 ::pc/resolver-dispatch resolver-fn
-                 ::pc/indexes           @indexes})
-              pp/profile-plugin]}))
+  (p/parallel-parser
+    {::p/env     {::p/reader               [p/map-reader
+                                            pc/parallel-reader
+                                            pc/open-ident-reader
+                                            p/env-placeholder-reader]
+                  ::p/placeholder-prefixes #{">"}}
+     ::p/mutate  pc/mutate-async
+     ::p/plugins [(pc/connect-plugin {::pc/register latest-product})
+                  p/error-handler-plugin
+                  p/request-cache-plugin
+                  p/trace-plugin]}))
 
 (comment
   (parser {} [::latest-product]))