+++ src-docs/com/wsscode/pathom/book/connect/mutation_async.cljs
@@ -5,56 +5,45 @@
             [com.wsscode.pathom.connect :as pc]
             [com.wsscode.pathom.core :as p]))
 
-; setup indexes atom
-(def indexes (atom {}))
-
-; setup resolver dispatch and factory
-(defmulti resolver-fn pc/resolver-dispatch)
-(def defresolver (pc/resolver-factory resolver-fn indexes))
-
-; setup mutation dispatch and factory
-(defmulti mutation-fn pc/mutation-dispatch)
-(def defmutation (pc/mutation-factory mutation-fn indexes))
-
 (defn adapt-user [user]
   (-> (into {} (map (fn [[k v]] [(keyword "user" (name k)) v])) (dissoc user ::db/key))
       (assoc :user/id (::db/key user))))
 
-(defmutation 'user/create
-  {::pc/params [:user/name :user/email]
+(pc/defmutation user-create [{::keys [db]} user]
+  {::pc/sym    'user/create
+   ::pc/params [:user/name :user/email]
    ::pc/output [:user/id]}
-  (fn [{::keys [db]} user]
-    ; now we async to create the record into the indexeddb
-    (go
-      (let [db      (<? db)
-            user-id (-> user
-                        (select-keys [:user/name :user/email])
-                        (merge {:user/created-at (js/Date.)})
-                        (->> (db/create! {::db/db db ::db/store-name "users"}))
-                        <?)]
-        {:user/id       user-id
-         :app/id-remaps {(:user/id user) user-id}}))))
+  (go
+    (let [db      (<? db)
+          user-id (-> user
+                      (select-keys [:user/name :user/email])
+                      (merge {:user/created-at (js/Date.)})
+                      (->> (db/create! {::db/db db ::db/store-name "users"}))
+                      <?)]
+      {:user/id       user-id
+       :app/id-remaps {(:user/id user) user-id}})))
 
-(defresolver `user-by-id
+(pc/defresolver user-by-id [{::keys [db]} {:keys [user/id]}]
   {::pc/input  #{:user/id}
    ::pc/output [:user/id :user/name :user/email :user/created-at]}
-  (fn [{::keys [db]} {:keys [user/id]}]
-    (go
-      ; reading from indexeddb
-      (let [db (<? db)]
-        (-> (db/read-object {::db/db db ::db/store-name "users"} id) <?
-            adapt-user)))))
+  (go
+    ; reading from indexeddb
+    (let [db (<? db)]
+      (-> (db/read-object {::db/db db ::db/store-name "users"} id) <?
+          adapt-user))))
 
 ; let's make an access to all users
-(defresolver `all-users
+(pc/defresolver all-users [{::keys [db]} _]
   {::pc/output [{:user/all [:user/id :user/name :user/email :user/created-at]}]}
-  (fn [{::keys [db]} _]
-    (go
-      (let [db (<? db)]
-        (->> (db/scan-store {::db/db db ::db/store-name "users"})
-             (async/into []) <?
-             (mapv adapt-user)
-             (hash-map :user/all))))))
+  (go
+    (let [db (<? db)]
+      (->> (db/scan-store {::db/db db ::db/store-name "users"})
+           (async/into []) <?
+           (mapv adapt-user)
+           (hash-map :user/all)))))
+
+; list all our app resolvers and mutations
+(def app-registry [user-create user-by-id all-users])
 
 (def db-settings
   {::db/db-name    "connectAsyncDemo"
@@ -62,12 +51,16 @@
                                            ::db/indexes {"name" {::db/unique false}}}}}]})
 
 (def parser
-  (p/async-parser {::p/env     {::p/reader                 [p/map-reader pc/all-async-readers]
-                                ::pc/resolver-dispatch     resolver-fn
-                                ::pc/mutate-dispatch       mutation-fn
-                                ::pc/indexes               @indexes
-                                ::pc/mutation-join-globals [:app/id-remaps]
-                                ::db                       (db/setup-db db-settings)}
-                   ::p/plugins [p/error-handler-plugin]
-                   ::p/mutate  pc/mutate-async}))
-
+  (p/parallel-parser
+    {::p/env     {::p/reader               [p/map-reader
+                                            pc/parallel-reader
+                                            pc/open-ident-reader
+                                            p/env-placeholder-reader]
+                  ::p/placeholder-prefixes #{">"}
+                  ::pc/mutation-join-globals [:app/id-remaps]
+                  ::db                       (db/setup-db db-settings)}
+     ::p/mutate  pc/mutate-async
+     ::p/plugins [(pc/connect-plugin {::pc/register app-registry})
+                  p/error-handler-plugin
+                  p/request-cache-plugin
+                  p/trace-plugin]}))