+++ src-docs/com/wsscode/pathom/book/graphql/edn_converter.cljs
@@ -1,14 +1,17 @@
 (ns com.wsscode.pathom.book.graphql.edn-converter
   (:require
     [cljs.reader :refer [read-string]]
+    [cljs.pprint]
     [clojure.string :as str]
     [com.wsscode.pathom.book.app-types :as app-types]
     [com.wsscode.pathom.book.ui.codemirror :as codemirror]
     [com.wsscode.pathom.graphql :as gql]
     [fulcro.client :as fulcro]
-    [fulcro.client.dom :as dom]
+    [fulcro.client.localized-dom :as dom]
     [fulcro.client.mutations :as mutations]
     [fulcro.client.primitives :as fp]
+    [garden.selectors :as gs]
+    [com.wsscode.pathom.viz.codemirror :as pcm]
     [goog.object :as gobj]
     [goog.string :as gstr]))
 
@@ -41,19 +44,31 @@
                             :ui/om-next-query    "[]"
                             :ui/graphql-query    ""
                             :ui/translate-error? false}
-                           params))
+                      params))
    :ident         [::id ::id]
    :query         [::id :ui/om-next-query :ui/graphql-query :ui/translate-error?]
-   :css           [[:.container {:display       "grid" :grid-template-columns "50% 50%"
+   :css           [[:.container {:display       "grid" :grid-template-columns "1fr 20px 1fr"
                                  :margin-bottom "20px"}
                     [:pre {:margin "0"}]
-                    [:.CodeMirror {:height "100%"}]]
-                   [:.translate-error {:color "#f00"}]]}
-  (dom/div #js {:className (str (:container css) " " (if translate-error? (:translate-error css) ""))}
-    (codemirror/clojure {:value    om-next-query
-                         :onChange #(fp/transact! this `[(update-query {:ui/om-next-query ~%})])})
-    (codemirror/graphql {:value               graphql-query
-                         ::codemirror/options {::codemirror/readOnly true}})))
+                    [(gs/> "" :div) {:height   "500px"
+                                     :position "relative"}]
+                    [:.divisor-v {:position      "static"
+                                  :width         "20px"
+                                  :background    "#eee"
+                                  :border        "1px solid #e0e0e0"
+                                  :border-top    "0"
+                                  :border-bottom "0"
+                                  :z-index       "2"}]]
+                   [:.translate-error {:color "#f00"}]]
+   :css-include   [pcm/Editor]}
+  (dom/div {:classes [:.container (if translate-error? :.translate-error)]}
+    (codemirror/clojure {:force-index-update? true
+                         :value               om-next-query
+                         :onChange            #(fp/transact! this `[(update-query {:ui/om-next-query ~%})])})
+    (dom/div :.divisor-v)
+    (codemirror/graphql {:force-index-update? true
+                         :value               graphql-query
+                         ::pcm/options        {::pcm/readOnly true}})))
 
 (def graphql-query-translator (fp/factory GraphQlQueryTranslator))
 
@@ -64,33 +79,33 @@
   [this {::keys [translator query-examples]} _ css]
   {:initial-state (fn [params]
                     {::id             (random-uuid)
-                     ::query-examples '{"Simple"            [:user/id :user/name]
-                                        "Join"              [{:app/me [:user/id :user/name]}]
-                                        "Ident"             [{[:user/login "wilkerlucio"] [:bio :url]}
-                                                             {[:organization/login "clojure"] [:name :url]}
-                                                             {[:github.repository/name-and-owner ["pathom" "wilkerlucio"]] [:id :name]}]
-                                        "Join Parameters"   [{(:app/all-users {:limit 10})
-                                                              [:user/id :user/name]}]
-                                        "Enums"             [{:viewer
-                                                              [{(:starred-repositories {:first    10
-                                                                                        :order-by {:field     STARRED_AT
-                                                                                                   :direction DESC}})
-                                                                [{:nodes
-                                                                  [:id :name :updated-at]}]}]}]
-                                        "Union"             [{:app/timeline
-                                                              {:app/User     [:user/id :user/name]
-                                                               :app/Activity [:activity/id :activity/title
-                                                                              {:activity/user
-                                                                               [:user/name]}]}}]
-                                        "Inline Union"      '[{:app/timeline
-                                                               [:entity/id
-                                                                (:user/name {::gql/on :app/User})
-                                                                {(:activity/user {::gql/on :app/User})
-                                                                 [:user/email]}]}]
-                                        "Recursive queries" [{:type
-                                                              [:kind :name {:ofType 3}]}]
-                                        "Mutation"          [{(users/create {:user/id 123 :user/name "Foo"})
-                                                              [:client-mutation-id]}]}
+                     ::query-examples '[["Simple" [:user/id :user/name]]
+                                        ["Join" [{:app/me [:user/id :user/name]}]]
+                                        ["Ident" [{[:user/login "wilkerlucio"] [:bio :url]}
+                                                  {[:organization/login "clojure"] [:name :url]}
+                                                  {[:github.repository/name-and-owner ["pathom" "wilkerlucio"]] [:id :name]}]]
+                                        ["Join Parameters" [{(:app/all-users {:limit 10})
+                                                             [:user/id :user/name]}]]
+                                        ["Enums" [{:viewer
+                                                   [{(:starred-repositories {:first    10
+                                                                             :order-by {:field     STARRED_AT
+                                                                                        :direction DESC}})
+                                                     [{:nodes
+                                                       [:id :name :updated-at]}]}]}]]
+                                        ["Union" [{:app/timeline
+                                                   {:app/User     [:user/id :user/name]
+                                                    :app/Activity [:activity/id :activity/title
+                                                                   {:activity/user
+                                                                    [:user/name]}]}}]]
+                                        ["Inline Union" '[{:app/timeline
+                                                           [:entity/id
+                                                            (:user/name {::gql/on :app/User})
+                                                            {(:activity/user {::gql/on :app/User})
+                                                             [:user/email]}]}]]
+                                        ["Recursive queries" [{:type
+                                                               [:kind :name {:ofType 3}]}]]
+                                        ["Mutation" [{(users/create {:user/id 123 :user/name "Foo"})
+                                                      [:client-mutation-id]}]]]
                      ::translator     (fp/get-initial-state GraphQlQueryTranslator params)})
    :ident         [::id ::id]
    :query         [::id ::query-examples