+++ src/com/wsscode/common/async_clj.clj
@@ -49,6 +49,17 @@
        (let [~name res#]
          ~@body))))
 
+(defmacro let-chan*
+  "Like let-chan, but async errors will be returned instead of propagated"
+  [[name value] & body]
+  `(let [res# ~value]
+     (if (chan? res#)
+       (go-catch
+         (let [~name (async/<! res#)]
+           ~@body))
+       (let [~name res#]
+         ~@body))))
+
 (defmacro go-promise [& body]
   `(let [ch# (async/promise-chan)]
      (async/go