+++ src/com/wsscode/common/async_cljs.clj
@@ -36,6 +36,17 @@
        (let [~name res#]
          ~@body))))
 
+(defmacro let-chan*
+  "Handles a possible channel on value."
+  [[name value] & body]
+  `(let [res# ~value]
+     (if (chan? res#)
+       (go-catch
+         (let [~name (cljs.core.async/<! res#)]
+           ~@body))
+       (let [~name res#]
+         ~@body))))
+
 (defmacro go-promise [& body]
   `(let [ch# (cljs.core.async/promise-chan)]
      (async/go